Split up list tests
